+++ packages/language-support/src/tests/formatting/comments.test.ts
@@ -48,19 +48,11 @@ RETURN n`;
   });
 
   test('weird inline comments', () => {
-    const inlinemultiline = `MERGE (n) /* Ensuring the node exists */ 
+    const inlinemultiline = `MERGE /* Ensuring the node exists */ (n)
   ON CREATE SET n.prop = 0 /* Set default property */
-MERGE (a:A) /* Create or match 'a:A' */
-  -[:T]-> (b:B) /* Link 'a' to 'b' */
-RETURN a.prop /* Return the property of 'a' */
-`;
-    const expected = `MERGE (n) /* Ensuring the node exists */
-  ON CREATE SET n.prop = 0 /* Set default property */
-MERGE
-  (a:A)- /* Create or match 'a:A' */
-    [:T]->
-  (b:B) /* Link 'a' to 'b' */
+MERGE /* Create or match 'a:A' */ (a:A)-[:T]->(b:B) // test
 RETURN a.prop /* Return the property of 'a' */`;
+    const expected = inlinemultiline;
     verifyFormatting(inlinemultiline, expected);
   });
 
@@ -302,12 +294,8 @@ RETURN n`;
     const query = `
 MATCH (a:Node) // first match
 WITH a, /* intermediate comment */ a.property AS prop
-RETURN prop; // final return`;
-    const expected = `MATCH (a:Node) // first match
-WITH
-  a, /* intermediate comment */
-  a.property AS prop
-RETURN prop; // final return`;
+RETURN prop; // final return`.trimStart();
+    const expected = query;
     verifyFormatting(query, expected);
   });
 
@@ -553,6 +541,8 @@ WITH // Passes results to the next clause
   p,
   q
 UNWIND // Expands lists into multiple rows
+  [1, 2, 3] AS num;
+RETURN // Expands lists into multiple rows
   [1, 2, 3] AS num;`.trimStart();
     const expected = query.trim();
     verifyFormatting(query, expected);
@@ -564,6 +554,65 @@ MATCH // One comment.
   // Another comment. Really?
   (m)-[:RELATION]->(n)
 RETURN m, n`.trim();
+    const expected = query;
+    verifyFormatting(query, expected);
+  });
+
+  test('inline comment', () => {
+    const query = `
+MATCH /* One comment. */ (m)-[:RELATION]->(n)
+RETURN m, n`.trim();
+    const expected = query;
+    verifyFormatting(query, expected);
+  });
+
+  test('inline comment within a chunk', () => {
+    const query = `MATCH (m/*comment*/)-[:RELATION]->(n)
+RETURN m, n`;
+    const expected = `MATCH (m)- /*comment*/ [:RELATION]->(n)
+RETURN m, n`;
+    verifyFormatting(query, expected);
+  });
+
+  test('long inline comment within a chunk', () => {
+    const query = `MATCH (m/*commentcommentcommentcommentcommentcommentcommentcom*/)-[:RELATION]->(n)
+RETURN m, n`;
+    const expected = `MATCH
+  (m)- /*commentcommentcommentcommentcommentcommentcommentcom*/ [:RELATION]->(n)
+RETURN m, n`;
+    verifyFormatting(query, expected);
+  });
+
+  test('long inline comment within a chunk should break everything', () => {
+    // See the extra m at the end of the comment compared to above test, that m makes it over 80 characters
+    const query = `MATCH (m/*commentcommentcommentcommentcommentcommentcommentcomm*/)-[:RELATION]->(n)
+RETURN m, n`;
+    const expected = `MATCH
+  (m)- /*commentcommentcommentcommentcommentcommentcommentcomm*/ [:RELATION]->
+  (n)
+RETURN m, n`;
+    verifyFormatting(query, expected);
+  });
+
+  test('long list in a return with comment', () => {
+    const query = `
+RETURN // test
+  [
+    "OCj0AswA",
+    "dFRbj1s3",
+    "oMbdvgm7",
+    "L4Vey8xn",
+    "GNgeDIkA",
+    "pU4RE0lM",
+    "M6XNVJsO",
+    "NcdW0tuB",
+    "Pf6RIuP4",
+    "6tKStKwl",
+    "HfvahDu5",
+    "gJoq3HnU",
+    "g7LjxbGD"
+  ]
+RETURN p`.trimStart();
     const expected = query;
     verifyFormatting(query, expected);
   });